backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-adk
openai-whisper
pyttsx3
pywin32
pydub
python-multipart
torch
scipy
chromadb
sentence-transformers
pypdf
